1) Upgraded System.Data.SQLite source from 1.0.66.0 to 1.0.84.0 (which also includes a base SQLite3 upgrade)
	Compiled with flag to enable foreign key enforcement by default
	New feature we needed: ability to specify enforcement of foreign keys via connection string key\value pair so we can turn it of when creating circularly referenced tables but have it on for everything else for data integrity purposes
	This time, I recorded every last change made to the source solutions\projects\code in sqlite-netFx-source-1.0.84.0.changes.txt
	Replaced System.Data.SQLite 3rdParty binaries with these new versions, both 32-bit and 64-bit
2) Created SQLProxies classes\interfaces to allow wrapping the architecture-specific binaries for use in an ANY CPU (MSIL) architecture C# library
	Odd note: Exposing the properties ErrorCode and ReturnCode in SQLiteException class revealed a critical error in the C# compiler and intermediate language, but a workaround is in place using int to int casting (int myInt = (int)[int property getter here]) and ambiguous property name detection to get a reflected PropertyInfo for a getter
3) Created new database structure including all new scripts for all tables and indexes, sql scripts are in SQLIndexer\SQLGenerationScripts (unencrypted and not included in project) and SQLIndexer\IndexDBScripts (encrypted and included as embedded resources); a diagram of the database can be found in Visio diagram in root of repository as SQLiteDB.vsd
4) Wrote new InitializeDatabase code which is called in SQLIndexer MonitorAgent CreateAndInitialize method
5) Adapted SQLAccessor CRUD operations class to use SQLite instead of SQL CE
6) Added CompilerServices in base of CloudApiPublic to allow the .NET 4.5 functionality in .NET 4 for a couple input parameter attributes: CallerMemberName and CallerFileLineNumber; will not work correctly at runtime if compiled in any Visual Studio before 2012

+++ CloudApiPublic/SQLIndexer/SqlModel/SqlEnum.cs
@@ -16,10 +16,10 @@ namespace Cloud.SQLIndexer.SqlModel
     internal class SqlEnum
     {
         [SqlAccess.Property]
-        public int EnumId { get; set; }
+        public long EnumId { get; set; }
 
         [SqlAccess.Property]
-        public int EnumCategoryId { get; set; }
+        public long EnumCategoryId { get; set; }
 
         [SqlAccess.Property]
         public string Name { get; set; }
